+++ src/Commands/0_core/about.js
@@ -1,6 +1,6 @@
 const Command = require('../../Structures/Command');
 
-const { MessageEmbed } = require('discord.js');
+const { EmbedBuilder } = require('discord.js');
 const { bot: { prefix, ownerID } } = require('../../../config/config.json');
 const { version, homepage, developerpage } = require('../../../package.json');
 
@@ -10,7 +10,7 @@ module.exports = new Command({
 	aliases: [ ' ' ],
 	description: 'Shows information about the bot.',
 	async run(message, args, client) {
-		const embed = new MessageEmbed()
+		const embed = new EmbedBuilder()
 			.setColor(0x3399FF)
 			.setAuthor({
 				name: `All about ${client.user.username}!`,