+++ src/Commands/1_music/leavecleanup.js
@@ -1,6 +1,6 @@
 const Command = require('../../Structures/Command');
 
-const { Permissions } = require('discord.js');
+const { PermissionsBitField } = require('discord.js');
 
 const queue = require('../../Data/queue');
 const { bot: { ownerID }, emoji: { success, warning, error }, response: { invalidPermissions, wrongChannel, noMusic } } = require('../../../config/config.json');
@@ -17,7 +17,7 @@ module.exports = new Command({
         if (!message.member.voice.channel || guildQueue.connection.joinConfig.channelId != message.member.voice.channel.id) return message.channel.send(`${warning} ${wrongChannel}`);
 
 		// check if message author has admin permissions
-		if (!message.member.permissions.has(Permissions.FLAGS.MANAGE_CHANNELS) && message.author.id != ownerID) return message.channel.send(`${error} ${invalidPermissions}`);
+		if (!message.member.permissions.has(PermissionsBitField.Flags.ManageChannels) && message.author.id != ownerID) return message.channel.send(`${error} ${invalidPermissions}`);
 
 		// get list of users in voice channel
 		const voiceChannelUsers = message.guild.channels.cache.get(guildQueue.connection.joinConfig.channelId).members.map(member => member.id);