+++ src/Commands/2_admin/forceprune.js
@@ -1,6 +1,6 @@
 const Command = require('../../Structures/Command.js');
 
-const { Permissions, ReactionCollector } = require('discord.js');
+const { PermissionsBitField, ReactionCollector } = require('discord.js');
 const { bot: { ownerID }, emoji: { success, warning }, response: { missingArguments, invalidPermissions, invalidNumber } } = require('../../../config/config.json');
 
 
@@ -12,7 +12,7 @@ module.exports = new Command({
     syntax: 'forceprune <amount> <confirm>',
 	description: 'Deletes the amount of messages **!ALL MESSAGES!** Requires Administrator permission.',
 	async run(message, args, client) {
-		if (!message.member.permissions.has(Permissions.FLAGS.ADMINISTRATOR) && message.author.id != ownerID) return message.channel.send(`${warning} ${invalidPermissions} (Administrator)`);
+		if (!message.member.permissions.has(PermissionsBitField.Flags.Administrator) && message.author.id != ownerID) return message.channel.send(`${warning} ${invalidPermissions} (Administrator)`);
         if (!args[0]) return message.channel.send(`${warning} ${missingArguments}`);
         if (isNaN(args[0])) return message.channel.send(`${warning} ${invalidNumber}`);
         if (args[0] > 99 || args[0] < 1) return message.channel.send(`${warning} Outside of number range!`);
@@ -41,8 +41,9 @@ module.exports = new Command({
                     result.delete(result.firstKey());  //remove the response message from the bulk delete
                     message.channel.bulkDelete(result);
 
-                    if (response.editable) response.edit(`${success} Deleting ${args[0]} messages`);
-                    setTimeout(() => response.delete(), 3750);
+                    if ( (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.edit(`${success} Deleting ${args[0]} messages`);
+                    await new Promise(resolve => setTimeout(resolve, 3750));
+                    if ( (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.delete();
 
                     collector.stop();
                     break;
@@ -51,7 +52,7 @@ module.exports = new Command({
                     //wait for allEMoji to be true
                     collector.stop();
                     while (!allEmoji) await new Promise(resolve => setTimeout(resolve, 100));
-                    if (response.deletable) response.reactions.removeAll();
+                    if ( (await response.channel.messages.fetch({ limit: 1, cache: false, around: response.id })).has(response.id) ) response.reactions.removeAll();
                     break;
             }
         });